+++ application/app/services/repo/repo_resolver_context.rb
@@ -21,21 +21,4 @@ module Repo
       RepoResolverResponse.new(object_url, type)
     end
   end
-
-  class RepoResolverResponse
-    attr_reader :object_url, :type
-    def initialize(object_url, type)
-      @object_url = object_url
-      @type = type
-    end
-
-    def resolved?
-      object_url.present? && type.present?
-    end
-
-    def unknown?
-      type.nil?
-    end
-  end
 end
-
